requirements.txt

fastapi
uvicorn
pyswisseph
